+++ Modules/Microsoft365DSC/Modules/M365DSCUtil.psm1
@@ -2614,6 +2614,9 @@ function Test-M365DSCDependenciesForNewVersions
         }
         $i++
     }
+    
+    # The progress bar seems to hang sometimes. Make sure it is no longer displayed.
+    Write-Progress -Activity 'Scanning Dependencies' -Completed
 }
 
 <#
@@ -2688,6 +2691,9 @@ function Update-M365DSCDependencies
         }
         $i++
     }
+    
+    # The progress bar seems to hang sometimes. Make sure it is no longer displayed.
+    Write-Progress -Activity 'Scanning Dependencies' -Completed
 
     if ($ValidateOnly)
     {